+++ R/demoRepo.R
@@ -40,6 +40,8 @@ demoRepo <- function(.project_name) {
   
   # Add scripts to the repo
   fs::dir_create("script/pk", recurse = TRUE)
+  fs::dir_create("deliv/figure")
+  fs::dir_create("deliv/table")
   
   writeLines(
     c(
@@ -132,6 +134,112 @@ demoRepo <- function(.project_name) {
       '- run `logPending()` to see what scripts are in need of QC',
       '- use `logAccept()` to sign off on any scripts with pending QC'),
     "README.md"
+  )
+  
+  grDevices::pdf("deliv/figure/example-pdf1.pdf", width = 8, height = 11)
+  plot(1:10)
+  grDevices::dev.off()
+  
+  grDevices::png("deliv/figure/example-png1.png")
+  plot(1:10)
+  grDevices::dev.off()
+  
+  grDevices::pdf("deliv/figure/example-pdf2.pdf", width = 11, height = 8)
+  plot(5:10)
+  plot(1:1000)
+  grDevices::dev.off()
+  
+  grDevices::pdf("deliv/figure/example-pdf3.pdf")
+  plot(1:10)
+  grDevices::dev.off()
+  
+  pmdata <- pmtables::stdata()
+  
+  pmtables::stable_save(
+    x = pmtables::stable(pmdata, panel = "STUDY"),
+    file = "example-table-1.tex",
+    dir = "deliv/table"
+  )
+  
+  pmtables::stable_save(
+    x = pmtables::stable_long(
+      dplyr::bind_rows(
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata
+      )
+    ),
+    file = "example-table-long-1.tex",
+    dir = "deliv/table"
+  )
+  
+  system("svn add 'deliv/figure/example-pdf1.pdf' -q -q")
+  system("svn add 'deliv/figure/example-png1.png' -q -q")
+  system("svn add 'deliv/figure/example-pdf2.pdf' -q -q")
+  system("svn add 'deliv/figure/example-pdf3.pdf' -q -q")
+  system("svn add 'deliv/table/example-table-1.tex' -q -q")
+  system("svn add 'deliv/table/example-table-long-1.tex' -q -q")
+  system(glue::glue("svn commit -m 'add pdf' -q -q"))
+  
+  Sys.sleep(1)
+  
+  grDevices::pdf("deliv/figure/example-pdf1.pdf", width = 8, height = 11)
+  plot(5:1000)
+  grDevices::dev.off()
+  
+  grDevices::png("deliv/figure/example-png1.png")
+  plot(5:1000)
+  grDevices::dev.off()
+  
+  Sys.sleep(1)
+  
+  grDevices::pdf("deliv/figure/example-pdf2.pdf", width = 11, height = 8)
+  plot(1:2)
+  plot(1:4)
+  grDevices::dev.off()
+  
+  Sys.sleep(1)
+  
+  grDevices::pdf("deliv/figure/example-pdf3.pdf")
+  plot(1:100)
+  plot(1:3)
+  grDevices::dev.off()
+  
+  grDevices::pdf("deliv/figure/example-pdf4.pdf")
+  plot(1:300)
+  grDevices::dev.off()
+  
+  # Make an edit to the table
+  pmdata$N[1] <- "81"
+  
+  pmtables::stable_save(
+    x = pmtables::stable(pmdata),
+    file = "example-table-1.tex", 
+    dir = "deliv/table"
+  )
+  
+  pmtables::stable_save(
+    x = pmtables::stable_long(
+      dplyr::bind_rows(
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata,
+        pmdata
+      )
+    ),
+    file = "example-table-long-1.tex",
+    dir = "deliv/table"
   )
   
   repoDir